refactor: move client and server stats into pgdog-stats
