+++ pgdog-stats/src/lib.rs
@@ -1,5 +1,10 @@
+pub mod client;
+pub mod memory;
 pub mod pool;
 pub mod replication;
+pub mod server;
+pub mod state;
 
+pub use memory::*;
 pub use pool::*;
 pub use replication::*;
