+++ pgdog/src/backend/pool/config.rs
@@ -5,10 +5,10 @@ use std::{
     time::Duration,
 };
 
-use pgdog_config::{pooling::ConnectionRecovery, Role};
+use pgdog_config::Role;
 use serde::{Deserialize, Serialize};
 
-use crate::config::{Database, General, PoolerMode, User};
+use crate::config::{Database, General, User};
 
 /// Pool configuration.
 #[derive(Debug, Serialize, Deserialize, Clone, Copy, PartialEq)]
